fix(bug): and code optimization

+++ gen3/bin/gitops.sh
@@ -66,54 +66,60 @@ gen3_run_tfplan() {
 
   local message
   local slack_hook
+  local module
+  local changes
+  
   slack_hook="arn:aws:sns:us-east-1:433568766270:planx-csoc-alerts-topic"
-  if [ $1 == "vpc" ];
+  module=$1
+  changes=$(_check_cloud-automation_changes)
+
+  if [[ ${changes} == "true" ]];
   then
-    message=gen3_run_tfplan_vpc
-  elif [ $1 == "eks" ];
+    local changes
+    changes="$(git diff-index --name-only HEAD --)"
+    message=$(mktemp -p "$XDG_RUNTIME_DIR" "tmp_plan.XXXXXX")
+    echo "${vpc_name} has uncommited changes for cloud-automation" > ${message}
+    git diff-index --name-only HEAD -- >> ${message}
+  elif [[ ${changes} == "false" ]];
   then
-    message=gen3_run_tfplan_eks
+    # checking for the result of _check_cloud-automation_changes just in case it came out empty
+    # for whatever reson 
+    case "$module" in
+      "vpc")
+        message=$(_gen3_run_tfplan_vpc)
+        ;;
+      "eks")
+        message=$(_gen3_run_tfplan_eks)
+        ;;
+    esac
   fi
-  aws sns publish --target-arn ${slack_hook} --message file://${message} > /dev/null 2>&1
-  rm ${message}
+
+  if [ -n "${message}" ];
+  then
+    aws sns publish --target-arn ${slack_hook} --message file://${message} > /dev/null #2>&1
+    rm ${message}
+  fi
+
 }
 
 #
 # Function that checks for uncomitter changes to cloud-automation
 # and also if there are unapplied changes to the vpc module
 #
-gen3_run_tfplan_vpc() {
+_gen3_run_tfplan_vpc() {
 
   local plan
   local slack_hook
   local tempFile
-  slack_hook="arn:aws:sns:us-east-1:433568766270:planx-csoc-alerts-topic"
-  if [[ $(_check_cloud-automation_changes) == "false" ]];
-  then
-    gen3 workon $(grep profile ~/.aws/config  |awk '{print $2}'| cut -d] -f1) ${vpc_name} > /dev/null 2>&1
-    plan=$(gen3 tfplan | grep "Plan")
 
-    if [ -n "${plan}" ];
-    then
-      #echo ${plan}
-      tempFile=$(mktemp -p "$XDG_RUNTIME_DIR" "tmp_plan.XXXXXX")
-      echo "${vpc_name} has unapplied plan:" > ${tempFile}
-      echo -e "${plan}"| sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[mGK]//g" >> ${tempFile}
-      #echo ${tempFile}
-      #aws sns publish --target-arn ${slack_hook} --message file://${tempFile} > /dev/null 2>&1
-      #  "${vpc_name} has unapplied plan \n${plan}"
-      #rm ${tempFile}
-    fi
-  else
-    local changes
-    changes="$(git diff-index --name-only HEAD --)"
+  gen3 workon $(grep profile ~/.aws/config  |awk '{print $2}'| cut -d] -f1) ${vpc_name} > /dev/null 2>&1
+  plan=$(gen3 tfplan | grep "Plan")
+
+  if [ -n "${plan}" ];
+  then
     tempFile=$(mktemp -p "$XDG_RUNTIME_DIR" "tmp_plan.XXXXXX")
-    echo "${vpc_name} has uncommited changes for cloud-automation" > ${tempFile}
-    git diff-index --name-only HEAD -- >> ${tempFile}
-    #echo ${tempFile}
-    #aws sns publish --target-arn ${slack_hook} --message file://${tempFile} > /dev/null 2>&1
-    # "${vpc_name} has uncommited changes for cloud-automation \n${changes}"
-    #rm ${tempFile}
+    echo "${vpc_name} has unapplied plan:" > ${tempFile}
+    echo -e "${plan}"| sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[mGK]//g" >> ${tempFile}
   fi
   echo "${tempFile}"
 }
@@ -122,37 +128,20 @@ gen3_run_tfplan_vpc() {
 # Function that checks for uncomitter changes to cloud-automation
 # and also if there are unapplied changes to the eks module
 #
-gen3_run_tfplan_eks() {
+_gen3_run_tfplan_eks() {
 
   local plan
   local slack_hook
   local tempFile
-  slack_hook="arn:aws:sns:us-east-1:433568766270:planx-csoc-alerts-topic"
-  if [[ $(_check_cloud-automation_changes) == "false" ]];
-  then
-    gen3 workon $(grep profile ~/.aws/config  |awk '{print $2}'| cut -d] -f1) ${vpc_name}_eks > /dev/null 2>&1
-    plan=$(gen3 tfplan | grep "Plan")
 
-    if [ -n "${plan}" ];
-    then
-      #echo ${plan}
-      tempFile=$(mktemp -p "$XDG_RUNTIME_DIR" "tmp_plan.XXXXXX")
-      echo "${vpc_name}_eks has unapplied plan:" > ${tempFile}
-      echo -e "${plan}"| sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[mGK]//g" >> ${tempFile}
-      #echo ${tempFile}
-      #aws sns publish --target-arn ${slack_hook} --message file://${tempFile} > /dev/null 2>&1
-      #  "${vpc_name} has unapplied plan \n${plan}"
-      #rm ${tempFile}
-    fi
-  else
-    local changes
-    changes="$(git diff-index --name-only HEAD --)"
+  gen3 workon $(grep profile ~/.aws/config  |awk '{print $2}'| cut -d] -f1) ${vpc_name}_eks > /dev/null 2>&1
+  plan=$(gen3 tfplan | grep "Plan")
+
+  if [ -n "${plan}" ];
+  then
     tempFile=$(mktemp -p "$XDG_RUNTIME_DIR" "tmp_plan.XXXXXX")
-    echo "${vpc_name} has uncommited changes for cloud-automation" > ${tempFile}
-    git diff-index --name-only HEAD -- >> ${tempFile}
-    #aws sns publish --target-arn ${slack_hook} --message file://${tempFile} > /dev/null 2>&1
-    # "${vpc_name} has uncommited changes for cloud-automation \n${changes}"
-    #rm ${tempFile}
+    echo "${vpc_name}_eks has unapplied plan:" > ${tempFile}
+    echo -e "${plan}"| sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[mGK]//g" >> ${tempFile}
   fi
   echo "${tempFile}"
 }
@@ -765,12 +754,6 @@ if [[ -z "$GEN3_SOURCE_ONLY" ]]; then
       ;;
     "dotag")
       gen3_gitops_repo_dotag "$@"
-      ;;
-    "tfplan-vpc")
-      gen3_run_tfplan_vpc "$@"
-      ;;
-    "tfplan-eks")
-      gen3_run_tfplan_eks "$@"
       ;;
     "tfplan")
       gen3_run_tfplan "$@"
